+++ src/Adapter/Options/OptionsInterface.php
@@ -22,4 +22,30 @@ interface OptionsInterface
      * @return OptionsInterface
      */
     public function setOptions(array $options);
+
+    /**
+     * Set language
+     * @param string $language
+     * @return OptionsInterface
+     */
+    public function setLanguage($language);
+
+    /**
+     * Get language
+     * @return string
+     */
+    public function getLanguage();
+
+    /**
+     * Set voice
+     * @param string $voice
+     * @return OptionsInterface
+     */
+    public function setVoice($voice);
+
+    /**
+     * Get voice
+     * @return string
+     */
+    public function getVoice();
 }